requirements.txt
asgiref==3.5.0
Django==3.2.3
install==1.3.5
psycopg2==2.9.3
pytz==2021.3
sorl-thumbnail==12.7.0
sqlparse==0.4.2
suds-py3==1.4.5.0
var-dump==1.2
